refactor: [#1396] extract event module in HTTP core
