+++ packages/http-tracker-core/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod container;
+pub mod event;
 pub mod services;
 pub mod statistics;
