+++ packages/http-tracker-core/src/statistics/event/listener.rs
@@ -1,7 +1,7 @@
 use tokio::sync::broadcast;
 
 use super::handler::handle_event;
-use super::Event;
+use crate::event::Event;
 use crate::statistics::repository::Repository;
 
 pub async fn dispatch_events(mut receiver: broadcast::Receiver<Event>, stats_repository: Repository) {
